docs/requirements.txt
matplotlib==3.5.1
numpydoc==1.2
pydata-sphinx-theme
Sphinx
Pillow